requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.10.0
attrs==25.3.0
billiard==4.2.2
boto3==1.40.48
botocore==1.40.48
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dj-database-url==3.1.2
Django==4.2.7
django-cors-headers==4.9.0
django-environ==0.12.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-spectacular==0.29.0
faiss-cpu==1.11.0
fastapi==0.129.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
greenlet==3.2.3
groq==0.29.0
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
ifaddr==0.2.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.5.4
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-deepseek==0.1.3
langchain-groq==0.3.5
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langsmith==0.4.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.3.1
openai==1.93.0
orjson==3.10.18
packaging==24.2
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
psycopg2-binary==2.9.10
py_eureka_client==0.13.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.11.0
pypdf==5.8.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.4.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.30.0
s3transfer==0.14.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
tree-sitter==0.24.0
tree-sitter-java==0.23.5
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
